+++ client_manager/src/main/java/com/simpolab/client_manager/session/SessionController.java
@@ -3,10 +3,7 @@ package com.simpolab.client_manager.session;
 import com.google.gson.Gson;
 import com.simpolab.client_manager.domain.Option;
 import com.simpolab.client_manager.domain.Session;
-import com.simpolab.client_manager.utils.ErrorBody;
-import com.simpolab.client_manager.utils.HttpUtils;
-import com.simpolab.client_manager.utils.JsonUtils;
-import com.simpolab.client_manager.utils.SceneUtils;
+import com.simpolab.client_manager.utils.*;
 import java.net.URL;
 import java.util.Date;
 import java.util.List;
@@ -43,6 +40,9 @@ public class SessionController implements Initializable {
 
   @FXML
   private Button btnStart;
+
+  @FXML
+  private Button btnDelete;
 
   @FXML
   private BarChart<Option, Integer> barChartVotes;
@@ -92,6 +92,7 @@ public class SessionController implements Initializable {
 
   @FXML
   private void onBtnDeleteClicked(ActionEvent event) throws Exception {
+    if (!AlertUtils.confirmAlert("Do you want to delete this session?")) return;
     HttpUtils.delete("/api/v1/session/" + session.getId());
     SceneUtils.switchTo("session/sessions.fxml");
   }
@@ -142,7 +143,7 @@ public class SessionController implements Initializable {
         }
         case 2 -> lblWinner.setText("Quorum has not been reached");
         case 3 -> lblWinner.setText("Absolute majority has not been reached");
-        default -> lblWinner.setText("wtf");
+        default -> lblWinner.setText("Unknown message");
       }
       return;
     }
@@ -195,6 +196,7 @@ public class SessionController implements Initializable {
 
     if (!session.getState().equals(Session.State.INACTIVE)) {
       btnStart.setDisable(true);
+      vboxContainer.getChildren().remove(btnDelete);
     }
 
     if (!session.getState().equals(Session.State.ACTIVE)) {
